+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 * add `useKeyMap`, `usePlayback`, `useScript` hooks
 
+* add `Playback.timeline` for much easier animation
+
 ## 2.0.5 (May 28, 2021)
 
 * correctly remove listeners when unmounting `<Audio>`/`<Video>`